Extract variants-related logic into a separate file
