+++ flow-components-parent/flow-webcomponent-api-analyzer/package.json
@@ -6,10 +6,10 @@
   "license": "Apache License 2.0",
   "dependencies": {
     "bower": "^1.8.2",
+    "cheerio": "^0.22.0",
     "coa": "~1.0.1",
     "fs-extra": "latest",
     "gulp": "latest",
-    "gulp-cheerio": "^0.6.3",
     "gulp-filter": "latest",
     "gulp-ignore": "latest",
     "minimist": "~1.1.1",
